requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
distro==1.9.0
fastapi==0.135.2
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.1
google-genai==1.68.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
hyperframe==6.1.0
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
litellm==1.82.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.3
openai==2.29.0
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
PyMuPDF==1.27.2.2
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
python-docx==1.2.0
python-dotenv==1.2.2
python-magic==0.4.27
python-multipart==0.0.22
python-telegram-bot==22.7
PyYAML==6.0.3
qdrant-client==1.17.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
sse-starlette==3.3.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
